[REFACTOR] 프로젝트 도메인 디렉토리 구조 변경
기능 별로 디렉토리 구분
