+++ team/src/main/java/sansam/team/project/command/application/controller/board/ProjectBoardController.java
@@ -1,15 +1,14 @@
-package sansam.team.project.command.application.controller;
+package sansam.team.project.command.application.controller.board;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import sansam.team.project.command.application.dto.board.ProjectApplyMemberDTO;
 import sansam.team.project.command.application.dto.board.ProjectBoardCreateDTO;
 import sansam.team.project.command.application.dto.board.ProjectBoardUpdateDTO;
-import sansam.team.project.command.application.service.ProjectBoardService;
+import sansam.team.project.command.application.service.board.ProjectBoardService;
 import sansam.team.project.command.domain.aggregate.entity.ProjectApplyMember;
 import sansam.team.project.command.domain.aggregate.entity.ProjectBoard;
