+++ team/src/main/java/sansam/team/project/command/application/service/board/ProjectApplyMemberService.java
@@ -1,4 +1,4 @@
-package sansam.team.project.command.application.service;
+package sansam.team.project.command.application.service.board;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
@@ -9,8 +9,8 @@ import sansam.team.common.jwt.SecurityUtil;
 import sansam.team.project.command.application.dto.board.ProjectApplyMemberDTO;
 import sansam.team.project.command.domain.aggregate.entity.ProjectApplyMember;
 import sansam.team.project.command.domain.aggregate.entity.ProjectBoard;
-import sansam.team.project.command.domain.repository.ProjectApplyMemberRepository;
-import sansam.team.project.command.domain.repository.ProjectBoardRepository;
+import sansam.team.project.command.domain.repository.board.ProjectApplyMemberRepository;
+import sansam.team.project.command.domain.repository.board.ProjectBoardRepository;
 import sansam.team.project.command.mapper.ProjectApplyMemberMapper;
 import sansam.team.user.command.domain.aggregate.entity.User;
