+++ team/src/main/java/sansam/team/project/command/infrastructure/repository/board/JpaProjectBoardRepository.java
@@ -1,7 +1,7 @@
-package sansam.team.project.command.infrastructure.repository;
+package sansam.team.project.command.infrastructure.repository.board;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sansam.team.project.command.domain.repository.ProjectBoardRepository;
+import sansam.team.project.command.domain.repository.board.ProjectBoardRepository;
 import sansam.team.project.command.domain.aggregate.entity.ProjectBoard;
 
 public interface JpaProjectBoardRepository extends ProjectBoardRepository, JpaRepository<ProjectBoard, Long> {
