+++ team/src/main/java/sansam/team/project/command/infrastructure/repository/project/JpaProjectRepository.java
@@ -1,7 +1,7 @@
-package sansam.team.project.command.infrastructure.repository;
+package sansam.team.project.command.infrastructure.repository.project;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sansam.team.project.command.domain.repository.ProjectRepository;
+import sansam.team.project.command.domain.repository.project.ProjectRepository;
 import sansam.team.project.command.domain.aggregate.entity.Project;
 
 public interface JpaProjectRepository extends ProjectRepository, JpaRepository<Project, Long> {
