+++ team/src/main/java/sansam/team/project/query/controller/project/ProjectAdminQueryController.java
@@ -1,5 +1,7 @@
-package sansam.team.project.query.controller;
+package sansam.team.project.query.controller.project;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,6 +17,7 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/v1/admin/project")
 @RequiredArgsConstructor
+@Tag(name = "Project Admin API", description = "프로젝트 관리자 API")
 public class ProjectAdminQueryController {
 
     private final ProjectQueryService projectQueryService;
@@ -23,6 +26,7 @@ public class ProjectAdminQueryController {
      * 프로젝트 전체 조회 (관리자용)
      */
     @GetMapping
+    @Operation(summary = "프로젝트 전체 조회", description = "프로젝트 전체 조회 API (관리자만 가능)")
     public ResponseEntity<List<ProjectAllQueryDTO>> getAllProjectsForAdmin() {
         List<ProjectAllQueryDTO> projects = projectQueryService.getAllProjectsForAdmin();
         return ResponseEntity.ok(projects);
@@ -32,6 +36,7 @@ public class ProjectAdminQueryController {
      * 프로젝트 상세 조회 (관리자용)
      */
     @GetMapping("/{projectSeq}")
+    @Operation(summary = "프로젝트 상세 조회", description = "프로젝트 상세 조회 API (관리자만 가능)")
     public ResponseEntity<ProjectAdminQueryDTO> getProjectByIdForAdmin(@PathVariable Long projectSeq) {
         ProjectAdminQueryDTO project = projectQueryService.getProjectByIdForAdmin(projectSeq);
         return ResponseEntity.ok(project);
